hotfix: remove deferred logging backoff delay
In multidomain logging, there is a backoff delay which is used to print
logging messages from different sources in the correct order. In the
current state of the sample on STM32H747, we never return from this
delay, and logging messages are never displayed. Simply disabling the
delay fixes the issue.

Signed-off-by: Celina Sophie Kalus <hello@celinakalus.de>